backend/requirements.txt
fastapi
uvicorn
python-multipart
open_clip_torch>=2.31.0
timm>=1.0.15
opencv-python
pillow
numpy
tqdm
torch
torchvision
transformers
tokenizers
scikit-learn==1.6.1
xgboost
git+https://github.com/facebookresearch/sam2.git